iOS 10 Beta 2

+++ Frameworks/AVFoundation.framework/AVAssetWriterFailedTerminalHelper.h
@@ -14,6 +14,6 @@
 - (id)initWithConfigurationState:(id)arg1;
 - (id)initWithConfigurationState:(id)arg1 terminalError:(id)arg2;
 - (void)startSessionAtSourceTime:(struct { long long x1; int x2; unsigned int x3; long long x4; })arg1;
-- (int)status;
+- (long long)status;
 
 @end
