+++ Frameworks/CFNetwork.framework/__NSCFURLLocalStreamTaskWorkRead.h
@@ -4,8 +4,8 @@
 
 @interface __NSCFURLLocalStreamTaskWorkRead : __NSCFURLLocalStreamTaskWork {
     id /* block */  _completion;
-    unsigned int  _maxBytes;
-    unsigned int  _minBytes;
+    unsigned long long  _maxBytes;
+    unsigned long long  _minBytes;
 }
 
 - (void)dealloc;
