+++ Frameworks/CoreImage.framework/CIRectangleDetector.h
@@ -3,9 +3,9 @@
  */
 
 @interface CIRectangleDetector : CIDetector {
-    float  _height;
+    double  _height;
     void * _internalBuffer;
-    float  _width;
+    double  _width;
     CIContext * context;
     NSMutableDictionary * featureOptions;
     struct __CVBuffer { } * interimScaleBuffer;
