+++ Frameworks/UIKit.framework/_UISearchControllerATVSearchBarAnimator.h
@@ -7,6 +7,6 @@
 - (void)animateTransition:(id)arg1;
 - (void)didFocusSearchBarForController:(id)arg1;
 - (void)didUnfocusSearchBarForController:(id)arg1;
-- (void)willFocusOffscreenViewForController:(id)arg1 withHeading:(unsigned int)arg2;
+- (void)willFocusOffscreenViewForController:(id)arg1 withHeading:(unsigned long long)arg2;
 
 @end
