+++ Frameworks/UIKit.framework/_UISearchControllerAnimator.h
@@ -6,13 +6,13 @@
 
 @property (readonly, copy) NSString *debugDescription;
 @property (readonly, copy) NSString *description;
-@property (readonly) unsigned int hash;
+@property (readonly) unsigned long long hash;
 @property (readonly) Class superclass;
 
 - (void)animateTransition:(id)arg1;
 - (void)didFocusSearchBarForController:(id)arg1;
 - (void)didUnfocusSearchBarForController:(id)arg1;
 - (double)transitionDuration:(id)arg1;
-- (void)willFocusOffscreenViewForController:(id)arg1 withHeading:(unsigned int)arg2;
+- (void)willFocusOffscreenViewForController:(id)arg1 withHeading:(unsigned long long)arg2;
 
 @end
