+++ Frameworks/UIKit.framework/_UITouchObservation.h
@@ -3,8 +3,8 @@
  */
 
 @interface _UITouchObservation : NSObject {
-    unsigned int  _observedTouchCount;
-    unsigned int  _observedTouchOffset;
+    unsigned long long  _observedTouchCount;
+    unsigned long long  _observedTouchOffset;
     UITouch * _observedTouches;
 }
 
@@ -13,6 +13,6 @@
 - (void)enumerateTouchesWithBlock:(id /* block */)arg1;
 - (void)observeTouch:(id)arg1;
 - (void)reset;
-- (unsigned int)touchCount;
+- (unsigned long long)touchCount;
 
 @end
