+++ PrivateFrameworks/CameraUI.framework/CAMRampToVideoZoomFactorCommand.h
@@ -4,18 +4,18 @@
 
 @interface CAMRampToVideoZoomFactorCommand : CAMCaptureCommand {
     float  __rate;
-    float  __videoZoomFactor;
+    double  __videoZoomFactor;
 }
 
 @property (nonatomic, readonly) float _rate;
-@property (nonatomic, readonly) float _videoZoomFactor;
+@property (nonatomic, readonly) double _videoZoomFactor;
 
 - (float)_rate;
-- (float)_videoZoomFactor;
+- (double)_videoZoomFactor;
 - (id)copyWithZone:(struct _NSZone { }*)arg1;
 - (void)encodeWithCoder:(id)arg1;
 - (void)executeWithContext:(id)arg1;
 - (id)initWithCoder:(id)arg1;
-- (id)initWithVideoZoomFactor:(float)arg1 rate:(float)arg2;
+- (id)initWithVideoZoomFactor:(double)arg1 rate:(float)arg2;
 
 @end
