+++ PrivateFrameworks/CarKit.framework/CRVehicleInformation.h
@@ -4,27 +4,27 @@
 
 @interface CRVehicleInformation : NSObject {
     AVExternalDevice * _externalDevice;
-    unsigned int  _mapsFallbackAmbientBrightness;
+    unsigned long long  _mapsFallbackAmbientBrightness;
 }
 
-@property (nonatomic, readonly) unsigned int driverPosition;
+@property (nonatomic, readonly) unsigned long long driverPosition;
 @property (nonatomic, retain) AVExternalDevice *externalDevice;
-@property (nonatomic, readonly) unsigned int interactionRestrictions;
-@property (nonatomic) unsigned int mapsFallbackAmbientBrightness;
-@property (nonatomic, readonly) unsigned int vehicleAmbientBrightness;
+@property (nonatomic, readonly) unsigned long long interactionRestrictions;
+@property (nonatomic) unsigned long long mapsFallbackAmbientBrightness;
+@property (nonatomic, readonly) unsigned long long vehicleAmbientBrightness;
 
 - (void).cxx_destruct;
 - (void)_handleLimitedUIChanged:(id)arg1;
 - (void)_handleNightModeChanged:(id)arg1;
 - (void)_screenDidUpdate:(id)arg1;
 - (void)dealloc;
-- (unsigned int)driverPosition;
+- (unsigned long long)driverPosition;
 - (id)externalDevice;
 - (id)init;
-- (unsigned int)interactionRestrictions;
-- (unsigned int)mapsFallbackAmbientBrightness;
+- (unsigned long long)interactionRestrictions;
+- (unsigned long long)mapsFallbackAmbientBrightness;
 - (void)setExternalDevice:(id)arg1;
-- (void)setMapsFallbackAmbientBrightness:(unsigned int)arg1;
-- (unsigned int)vehicleAmbientBrightness;
+- (void)setMapsFallbackAmbientBrightness:(unsigned long long)arg1;
+- (unsigned long long)vehicleAmbientBrightness;
 
 @end
