+++ PrivateFrameworks/CloudPhotoLibrary.framework/CPLChangeSession.h
@@ -7,21 +7,21 @@
     CPLPlatformObject * _platformObject;
     NSString * _sessionIdentifier;
     NSProgress * _sessionProgress;
-    unsigned int  _state;
+    unsigned long long  _state;
 }
 
 @property (readonly, copy) NSString *debugDescription;
 @property (readonly, copy) NSString *description;
-@property (readonly) unsigned int hash;
+@property (readonly) unsigned long long hash;
 @property (nonatomic, readonly) CPLLibraryManager *libraryManager;
 @property (nonatomic, readonly) CPLPlatformObject *platformObject;
 @property (nonatomic, readonly) NSString *sessionIdentifier;
 @property (nonatomic, readonly) NSProgress *sessionProgress;
-@property (nonatomic) unsigned int state;
+@property (nonatomic) unsigned long long state;
 @property (readonly) Class superclass;
 
 + (id)platformImplementationProtocol;
-+ (id)stateDescriptionForState:(unsigned int)arg1;
++ (id)stateDescriptionForState:(unsigned long long)arg1;
 
 - (void).cxx_destruct;
 - (id)_sessionLogDomain;
@@ -36,8 +36,8 @@
 - (void)resume;
 - (id)sessionIdentifier;
 - (id)sessionProgress;
-- (void)setState:(unsigned int)arg1;
-- (unsigned int)state;
+- (void)setState:(unsigned long long)arg1;
+- (unsigned long long)state;
 - (void)tearDownWithCompletionHandler:(id /* block */)arg1;
 
 @end
