+++ PrivateFrameworks/CompanionSync.framework/_SYInputStreamer.h
@@ -13,7 +13,7 @@
 
 @property (readonly, copy) NSString *debugDescription;
 @property (readonly, copy) NSString *description;
-@property (readonly) unsigned int hash;
+@property (readonly) unsigned long long hash;
 @property (readonly) Class superclass;
 
 - (void).cxx_destruct;
@@ -21,8 +21,8 @@
 - (void)_tryToReadData;
 - (void)close;
 - (id)initWithCompressedFileURL:(id)arg1 callbackQueue:(id)arg2;
-- (void)readDataOfLength:(unsigned int)arg1 completion:(id /* block */)arg2;
-- (void)stream:(id)arg1 handleEvent:(unsigned int)arg2;
+- (void)readDataOfLength:(unsigned long long)arg1 completion:(id /* block */)arg2;
+- (void)stream:(id)arg1 handleEvent:(unsigned long long)arg2;
 - (void)whenComplete:(id /* block */)arg1;
 
 @end
