+++ PrivateFrameworks/ContactsAutocomplete.framework/CNAutocompleteDuetSearch.h
@@ -10,17 +10,17 @@
 @property (nonatomic, readonly) CNContactStore *contactStore;
 @property (readonly, copy) NSString *debugDescription;
 @property (readonly, copy) NSString *description;
-@property (readonly) unsigned int hash;
+@property (readonly) unsigned long long hash;
 @property (nonatomic, readonly) _CDPeopleSuggester *peopleSuggester;
 @property (readonly) Class superclass;
 
 + (void)configureContext:(id)arg1 withFetchRequest:(id)arg2;
 + (void)configureSettings:(id)arg1 withFetchRequest:(id)arg2;
-+ (id)constrainMechanismsForSearchType:(unsigned int)arg1;
-+ (unsigned int)predictedResultLimit;
++ (id)constrainMechanismsForSearchType:(unsigned long long)arg1;
++ (unsigned long long)predictedResultLimit;
 
 - (void).cxx_destruct;
-- (int)addressTypeForCDContact:(id)arg1;
+- (long long)addressTypeForCDContact:(id)arg1;
 - (id)autocompleteResultsFromSuggesterResults:(id)arg1 factory:(id)arg2;
 - (id)contactStore;
 - (id)executeRequest:(id)arg1 completionHandler:(id /* block */)arg2;
