+++ PrivateFrameworks/ContactsFoundation.framework/_CNPriorityQueueingStrategy.h
@@ -8,10 +8,10 @@
 
 @property (readonly, copy) NSString *debugDescription;
 @property (readonly, copy) NSString *description;
-@property (readonly) unsigned int hash;
+@property (readonly) unsigned long long hash;
 @property (readonly) Class superclass;
 
-- (unsigned int)_insertionIndexForObject:(id)arg1 buffer:(id)arg2;
+- (unsigned long long)_insertionIndexForObject:(id)arg1 buffer:(id)arg2;
 - (void)dealloc;
 - (void)enqueueObject:(id)arg1 buffer:(id)arg2;
 - (id)initWithComparator:(id /* block */)arg1;
