+++ PrivateFrameworks/DataDetectorsUI.framework/DDTextMessageAction.h
@@ -6,12 +6,12 @@
 
 @property (readonly, copy) NSString *debugDescription;
 @property (readonly, copy) NSString *description;
-@property (readonly) unsigned int hash;
+@property (readonly) unsigned long long hash;
 @property (readonly) Class superclass;
 
 - (int)interactionType;
 - (id)localizedName;
-- (void)messageComposeViewController:(id)arg1 didFinishWithResult:(int)arg2;
+- (void)messageComposeViewController:(id)arg1 didFinishWithResult:(long long)arg2;
 - (void)perform;
 - (id)viewController;
