+++ PrivateFrameworks/DuetPLLConfigLogger.framework/AppLaunchStatsAggD.h
@@ -6,7 +6,7 @@
     int  BGReportToken;
     NSObject<OS_dispatch_queue> * aggDQueue;
     id /* block */  appisFetchable;
-    unsigned long  blameReasons;
+    unsigned long long  blameReasons;
     bool  buildLaunchStats;
     int  currentFetchFailuresIndex;
     NSDate * cycleStart;
@@ -62,7 +62,7 @@
 - (void)fetchAttemptIncrement;
 - (void)fetchCountUpdate:(int)arg1;
 - (void)fetchDataCount:(int)arg1;
-- (void)fetchDeniedBecause:(bool)arg1 withPWBudget:(bool)arg2 wifiQual:(int)arg3 cellQual:(int)arg4 networkingDeferred:(bool)arg5 ignorePoorCell:(bool)arg6 incall:(bool)arg7 thermal:(bool)arg8;
+- (void)fetchDeniedBecause:(bool)arg1 withPWBudget:(bool)arg2 wifiQual:(long long)arg3 cellQual:(long long)arg4 networkingDeferred:(bool)arg5 ignorePoorCell:(bool)arg6 incall:(bool)arg7 thermal:(bool)arg8;
 - (void)fetchDeniedForReason:(int)arg1;
 - (void)fetchDeniedForReasons:(id)arg1;
 - (void)fetchDeniesIncrement;
@@ -74,7 +74,7 @@
 - (void)fetchPWCount:(int)arg1;
 - (void)fetchTimeUpdate:(long long)arg1;
 - (void)forceReleaseCount;
-- (void)incrementTriggerReason:(int)arg1;
+- (void)incrementTriggerReason:(long long)arg1;
 - (id)init;
 - (void)newCycleWithApps:(id)arg1;
 - (int)numSlotsAfterPrewarmForApp:(id)arg1 launchedAt:(id)arg2;
