+++ PrivateFrameworks/FuseUI.framework/MusicMutableEntityPlaybackStatus.h
@@ -5,7 +5,7 @@
 @interface MusicMutableEntityPlaybackStatus : MusicEntityPlaybackStatus
 
 @property (nonatomic) double playbackCurrentTime;
-@property (nonatomic) int playbackDisplayState;
+@property (nonatomic) long long playbackDisplayState;
 @property (nonatomic) double playbackDuration;
 @property (nonatomic) float playbackRate;
 
@@ -13,7 +13,7 @@
 - (id)mutableCopyWithZone:(struct _NSZone { }*)arg1;
 - (void)setPlaybackCurrentTime:(double)arg1;
 - (void)setPlaybackCurrentTime:(double)arg1 originatingTime:(double)arg2;
-- (void)setPlaybackDisplayState:(int)arg1;
+- (void)setPlaybackDisplayState:(long long)arg1;
 - (void)setPlaybackDuration:(double)arg1;
 - (void)setPlaybackRate:(float)arg1;
