+++ PrivateFrameworks/FuseUI.framework/MusicNowPlayingAtmosphericEffectView.h
@@ -6,13 +6,13 @@
     _UIBackdropView * _backdropView;
     UIView * _blendModeTintView;
     UIView * _colorTintView;
-    int  _effect;
+    long long  _effect;
 }
 
 @property (nonatomic, retain) _UIBackdropView *backdropView;
 @property (nonatomic, retain) UIView *blendModeTintView;
 @property (nonatomic, retain) UIView *colorTintView;
-@property (nonatomic) int effect;
+@property (nonatomic) long long effect;
 
 + (Class)tintViewClass;
 
@@ -21,12 +21,12 @@
 - (id)backdropView;
 - (id)blendModeTintView;
 - (id)colorTintView;
-- (int)effect;
-- (id)initWithFrame:(struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })arg1;
+- (long long)effect;
+- (id)initWithFrame:(struct CGRect { struct CGPoint { double x_1_1_1; double x_1_1_2; } x1; struct CGSize { double x_2_1_1; double x_2_1_2; } x2; })arg1;
 - (void)layoutSubviews;
 - (void)setBackdropView:(id)arg1;
 - (void)setBlendModeTintView:(id)arg1;
 - (void)setColorTintView:(id)arg1;
-- (void)setEffect:(int)arg1;
+- (void)setEffect:(long long)arg1;
 
 @end
