+++ PrivateFrameworks/FuseUI.framework/MusicWelcomePlaceholderView.h
@@ -5,14 +5,14 @@
 @interface MusicWelcomePlaceholderView : UIView <MusicClientContextConsuming> {
     UIImageView * _backgroundImageView;
     MusicClientContext * _clientContext;
-    float  _logoImageTopInset;
+    double  _logoImageTopInset;
     UIImageView * _logoImageView;
 }
 
 @property (nonatomic, retain) SKUIClientContext *clientContext;
 @property (readonly, copy) NSString *debugDescription;
 @property (readonly, copy) NSString *description;
-@property (readonly) unsigned int hash;
+@property (readonly) unsigned long long hash;
 @property (readonly) Class superclass;
 
 - (void).cxx_destruct;
