+++ PrivateFrameworks/GeoServices.framework/GEORPProblemRequester.h
@@ -4,7 +4,7 @@
 
 @interface GEORPProblemRequester : GEOServiceRequester
 
-+ (unsigned int)_urlType;
++ (unsigned long long)_urlType;
 + (id)sharedInstance;
 
 - (id)_validateResponse:(id)arg1;
