+++ PrivateFrameworks/HealthDaemon.framework/HDSyncAnchorMap.h
@@ -9,11 +9,11 @@
 @property (nonatomic, retain) NSMutableDictionary *anchorsByObjectType;
 @property (readonly, copy) NSString *debugDescription;
 @property (readonly, copy) NSString *description;
-@property (readonly) unsigned int hash;
+@property (readonly) unsigned long long hash;
 @property (readonly) Class superclass;
 
 - (void).cxx_destruct;
-- (unsigned int)anchorCount;
+- (unsigned long long)anchorCount;
 - (long long)anchorForSyncEntityClass:(Class)arg1;
 - (id)anchorsByObjectType;
 - (id)description;
