+++ PrivateFrameworks/MetalTools.framework/MTLToolsPointerArray.h
@@ -3,7 +3,7 @@
  */
 
 @interface MTLToolsPointerArray : NSObject {
-    unsigned int  _pointerAdditions;
+    unsigned long long  _pointerAdditions;
     NSPointerArray * _pointerArray;
     NSObject<OS_dispatch_semaphore> * _semaphore;
 }
